+++ src/components/data-operators/serialize.js
@@ -1,0 +1,9 @@
+export const serialize = (reference) => {
+    let rawResult = {
+        content: null,
+        type: null,
+        length: null
+    };
+
+    return JSON.stringify(rawResult);
+};