+++ Changelog.md
@@ -3,3 +3,5 @@
 ### 1.1.0
 
 - Implemented `map()` function
+- Implemented `strictMap()` function
+- Implemented `strictForEach()` function